Init3

+++ ASP.NETCORE_PROJECT/ASP.NETCORE_PROJECT/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -274,8 +274,8 @@ namespace ASP.NETCORE_PROJECT.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("int");
+                    b.Property<double>("Price")
+                        .HasColumnType("float");
 
                     b.Property<int>("Quantity")
                         .HasColumnType("int");
